src/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
dotenv==0.9.9
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.1.0
packaging==24.2
postgrest==0.19.3
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
